OPIC-41 speed is probably quite optimal
2^17/0.006689
19595156.22664075347585588279
2^15/0.001389
23591072.71418286537077033837
2^20/0.146
7182027.39726027397260273972
2^25/7.16
4686373.18435754189944134078

When data is small, we can reach 20M up qps, which is 100x slower than cpu tick,
but that's expected.

When data is large, speed degrade to 4.xM qps, which is 1000x slower than cpu,
likely to be the memory access speed.  This is also expected because our data
cannot fit into cpu anymore.